pyproject.toml
[project]
name = "market-digest"
version = "0.1.0"
description = "Daily market report fetcher + summarizer agent"
requires-python = ">=3.12"
dependencies = [
    "requests>=2.32",
    "beautifulsoup4>=4.12",
    "lxml>=5.2",
    "pdfminer.six>=20240706",
    "pypdf>=4.3",
    "yfinance>=0.2.40",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "feedparser>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["market_digest"]
